Create methods for the different parts of the sortByUpvotedListings function and create a file for it. Also, create a folder for all of the Recommended Sorting related files.
